Move organism filters to filter folder
