+++ cg/store/api/find_basic_data.py
@@ -1,6 +1,6 @@
 """Handler to find basic data objects"""
 import datetime as dt
-from typing import List, Optional, Callable
+from typing import List, Optional
 
 from sqlalchemy import desc
 from sqlalchemy.orm import Query
@@ -17,7 +17,7 @@ from cg.store.models import (
     User,
 )
 from cg.store.api.base import BaseHandler
-from cg.store.organism_filters import OrganismFilter, apply_organism_filter
+from cg.store.filters.status_organism_filters import OrganismFilter, apply_organism_filter
 from cg.store.filters.status_bed_filters import apply_bed_filter, BedFilter
 from cg.store.filters.status_bed_version_filters import BedVersionFilter, apply_bed_version_filter
 from cg.store.filters.status_customer_filters import apply_customer_filter, CustomerFilter
