+++ tests/store/test_organism_filters.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Query
 
 from cg.store.api.core import Store
 from cg.store.models import Organism
-from cg.store.organism_filters import filter_organism_by_internal_id
+from cg.store.filters.status_organism_filters import filter_organism_by_internal_id
 
 
 def test_filter_organism_by_internal_id_returns_correct_organism(store_with_organisms: Store):
